+++ lab4_unit3/src/main/java/com/example/lab4_unit3/entity/InternalTask.java
@@ -1,4 +1,7 @@
 package com.example.lab4_unit3.entity;
 
+import jakarta.persistence.Entity;
+
+@Entity
 public class InternalTask extends Task{
 }
